pyproject.toml
[project]
name = "ops-job"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "channels>=4.3.1",
    "channels-redis>=4.3.0",
    "cryptography>=45.0.6",
    "django==5.0.1",
    "django-admin-autocomplete-filter>=0.7.1",
    "django-axes>=8.0.0",
    "django-cors-headers>=4.7.0",
    "django-db-connection-pool==1.2.5",
    "django-extensions>=3.2.3",
    "django-filter>=23.5",
    "django-guardian>=3.0.3",
    "django-json-widget>=2.0.3",
    "django-jsoneditor>=0.2.4",
    "django-simpleui>=2025.6.24",
    "djangorestframework>=3.14,<3.16",
    "djangorestframework-simplejwt>=5.3.0",
    "drf-extensions>=0.8.0",
    "drf-spectacular>=0.28.0",
    "paramiko>=4.0.0",
    "redis>=6.4.0",
    "croniter>=1.4.0",
    "psutil>=5.9.0",
    "python-json-logger>=3.3.0",
    "django-simple-captcha==0.6.2",
    "requests>=2.32.4",
    "django-health-check>=3.20.0",
    "factory-boy==3.3.0",
    "pytest-django==4.8.0",
    "pytest-cov==4.1.0",
    "django-redis>=6.0.0",
    "sentry-sdk==1.40.4",
    "python-dotenv>=1.0.0",
    "eventlet>=0.40.2",
    "daphne>=4.0.0",
    "django-debug-toolbar>=6.0.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "fabric>=3.2.2",
    "gevent>=25.8.2",
    "gunicorn>=22.0.0",
    "alibabacloud-ecs20140526>=7.2.1",
    "tencentcloud-sdk-python>=3.0.1458",
    #"django-auth-ldap>=5.0.0",
    #"python-ldap @ file:///${PROJECT_ROOT}/wheels/python_ldap-3.4.5-cp312-cp312-win_amd64.whl",
    #"python-ldap>=3.4.5",
    "django-otp>=1.6.3",
    "qrcode[pil]>=8.2",
    "django-apscheduler>=0.7.0",
    "djangorestframework-stubs>=3.16.6",
    "urlquote>=2.2.0",
    "psycopg2-binary>=2.9.9",
    "cos-python-sdk-v5>=1.9.40",
    "boto3>=1.42.14",
    "oss2>=2.19.1",
    "minio>=7.2.20",
    "websocket-client>=1.9.0",
    "tenacity>=9.1.2",
    "pyyaml>=6.0.2",
    "ruamel.yaml>=0.17.40,<0.18",
    "psutil>=5.9.0",
    "uvicorn>=0.40.0",
    "pytest-mock==3.12.0",
    "pytest-html==4.1.1",
    "pymysql>=1.1.2",
    "mysqlclient>=2.2.7",
]
